🚀🐛 Switches to `peaceiris/gh-pages` for Deploys
Due to a shortcoming of GitHub Actions[1], we are unable to use the in-built `GITHUB_TOKEN` secret in order to push to the `gh-pages` branch if we want it to in-turn trigger GitHub's internal process of publishing our changes. As such, we need to use a branch-pushing mechanism that uses a "Deploy" key such that it has full (GitHub-internal) permissions in order to trigger the downstream (internal) deployment process.

This switches us from using the [`gh-pages`](https://www.npmjs.com/package/gh-pages) NPM package to using the [`peaceiris/gh-pages`](https://github.com/peaceiris/actions-gh-pages) GitHub Action for these branch pushes, because the latter supports using the SSH key I've added as a secret to the repository to be used for deployments.

[1]: https://github.community/t5/GitHub-Actions/Github-action-not-triggering-gh-pages-upon-push/td-p/26869/highlight/true/page/2