🐛 Restores Handling of Content as Markdown
The content had previously been handled as markdown when it was retrieved through gatsby-transformer-remark. After the content started to be retrieved from gatsby-transformer-yaml, the content was no longer handled as markdown which led to some very subtle rendering differences (even though the content does not yet purposefully take advantage of markdown syntax.)

This restores the handling of the content as Markdown by directly consuming Remark, which fixes the subtle rendering issues and provides Markdown's plaintext-friendly formatting for content items should they need such in the future.